# src/main/java/rw/ac/rca/cloudsphere/identity/LoginRequest.java
package rw.ac.rca.cloudsphere.identity;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String password
) {}