# src/main/java/rw/ac/rca/cloudsphere/identity/RegisterRequest.java
package rw.ac.rca.cloudsphere.identity;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Email String email,
        @NotBlank @Size(min = 8, max = 72) String password,
        @NotBlank @Pattern(regexp = "^07[0-9]{8}$", message = "must be a 10-digit Rwandan mobile starting with 07")
        String phone
) {}